Adding routes and templates
