--- shifter/api/typesSOSProjects.go
@@ -1,10 +1,0 @@
-package api
-
-import (
-	osNativeProject "github.com/openshift/api/project/v1"
-)
-
-type SOSProjects struct {
-	Shifter  Shifter                     `json:"shifter"`
-	Projects osNativeProject.ProjectList `json:"projects"`
-}
