+++ shifter/openshift/deploymentConfig.go
@@ -18,18 +18,33 @@ import (
 	v1 "github.com/openshift/api/apps/v1"
 	appsv1 "github.com/openshift/client-go/apps/clientset/versioned"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"log"
 )
 
-func (c Openshift) GetDeploymentConfigs(namespace string) *v1.DeploymentConfigList {
-	app, _ := appsv1.NewForConfig(c.clusterClient())
-	depCfgLst, _ := app.AppsV1().DeploymentConfigs(namespace).List(context.TODO(), metav1.ListOptions{})
+func (c Openshift) GetAllDeploymentConfigs(namespace string) *v1.DeploymentConfigList {
+	app, err := appsv1.NewForConfig(c.clusterClient())
+	if err != nil {
+		log.Println(err)
+	}
+
+	depCfgLst, err := app.AppsV1().DeploymentConfigs(namespace).List(context.TODO(), metav1.ListOptions{})
+	if err != nil {
+		log.Println(err)
+	}
 
 	return depCfgLst
 }
 
 func (c Openshift) GetDeploymentConfig(name string, namespace string) *v1.DeploymentConfig {
-	app, _ := appsv1.NewForConfig(c.clusterClient())
-	depCfg, _ := app.AppsV1().DeploymentConfigs(namespace).Get(context.TODO(), name, metav1.GetOptions{})
+	app, err := appsv1.NewForConfig(c.clusterClient())
+	if err != nil {
+		log.Println(err)
+	}
+
+	depCfg, err := app.AppsV1().DeploymentConfigs(namespace).Get(context.TODO(), name, metav1.GetOptions{})
+	if err != nil {
+		log.Println(err)
+	}
 
 	return depCfg
 }
