+++ shifter/openshift/project.go
@@ -15,36 +15,36 @@ package openshift
 
 import (
 	"context"
-	v1 "github.com/openshift/api/route/v1"
-	routev1 "github.com/openshift/client-go/route/clientset/versioned"
+	v1 "github.com/openshift/api/project/v1"
+	projectv1 "github.com/openshift/client-go/project/clientset/versioned/typed/project/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"log"
 )
 
-func (c Openshift) GetRoutes(namespace string) *v1.RouteList {
-	routeCfg, err := routev1.NewForConfig(c.clusterClient())
+func (c Openshift) GetAllProjects() *v1.ProjectList {
+	projCfg, err := projectv1.NewForConfig(c.clusterClient())
 	if err != nil {
 		log.Println(err)
 	}
 
-	routeList, err := routeCfg.RouteV1().Routes(namespace).List(context.TODO(), metav1.ListOptions{})
+	projectList, err := projCfg.Projects().List(context.TODO(), metav1.ListOptions{})
 	if err != nil {
 		log.Println(err)
 	}
 
-	return routeList
+	return projectList
 }
 
-func (c Openshift) GetRoute(name string, namespace string) *v1.Route {
-	routeCfg, err := routev1.NewForConfig(c.clusterClient())
+func (c Openshift) GetProject(name string) *v1.Project {
+	projCfg, err := projectv1.NewForConfig(c.clusterClient())
 	if err != nil {
 		log.Println(err)
 	}
 
-	route, err := routeCfg.RouteV1().Routes(namespace).Get(context.TODO(), name, metav1.GetOptions{})
+	project, err := projCfg.Projects().Get(context.TODO(), name, metav1.GetOptions{})
 	if err != nil {
 		log.Println(err)
 	}
 
-	return route
+	return project
 }
