+++ shifter/openshift/route.go
@@ -1,0 +1,50 @@
+/*
+copyright 2019 google llc
+licensed under the apache license, version 2.0 (the "license");
+you may not use this file except in compliance with the license.
+you may obtain a copy of the license at
+    http://www.apache.org/licenses/license-2.0
+unless required by applicable law or agreed to in writing, software
+distributed under the license is distributed on an "as is" basis,
+without warranties or conditions of any kind, either express or implied.
+see the license for the specific language governing permissions and
+limitations under the license.
+*/
+
+package openshift
+
+import (
+	"context"
+	v1 "github.com/openshift/api/route/v1"
+	routev1 "github.com/openshift/client-go/route/clientset/versioned"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"log"
+)
+
+func (c Openshift) GetAllRoutes(namespace string) *v1.RouteList {
+	routeCfg, err := routev1.NewForConfig(c.clusterClient())
+	if err != nil {
+		log.Println(err)
+	}
+
+	routeList, err := routeCfg.RouteV1().Routes(namespace).List(context.TODO(), metav1.ListOptions{})
+	if err != nil {
+		log.Println(err)
+	}
+
+	return routeList
+}
+
+func (c Openshift) GetRoute(name string, namespace string) *v1.Route {
+	routeCfg, err := routev1.NewForConfig(c.clusterClient())
+	if err != nil {
+		log.Println(err)
+	}
+
+	route, err := routeCfg.RouteV1().Routes(namespace).Get(context.TODO(), name, metav1.GetOptions{})
+	if err != nil {
+		log.Println(err)
+	}
+
+	return route
+}
